completed backend

+++ backend/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc    Get goals
 // @route   GET /apis/goals
 // @access Private
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
 
     res.status(200).json(goals);
 })
@@ -26,6 +27,7 @@ const setGoal = asyncHandler(async (req, res) => {
     } else {
         const goal = await Goal.create({
             text: req.body.text,
+            user: req.user.id
         })
 
         console.log('Sending 200 response');
@@ -44,6 +46,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found');
     }
 
+    const user = await User.findById(req.user.id);
+
+    // Check for user
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Verify correct user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true});
     res.status(200).json(updatedGoal);
 })
@@ -58,6 +74,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400);
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // Check for user
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Verify correct user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const deletedGoal = await Goal.findOneAndDelete(req.params.id)
